fix: SearchBarを表示するかどうかを呼び出し側で行っていたところを、SearchBarコンポーネント内で行うように修正
