+++ src/app/dashboard-menu/header/search-bar/styles.module.css
@@ -1,0 +1,46 @@
+.root {
+  height: 92px;
+  padding: 0px 2.5rem;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  position: absolute;
+  background-color: rgba(249, 250, 251, 0.8);
+  top: 100%;
+  left: 0px;
+  box-shadow: rgba(145, 158, 171, 0.16) 0px 8px 16px 0px;
+  transition: transform 225ms cubic-bezier(0, 0, 0.2, 1);
+  transform: translateY(-100%);
+  margin-top: 0.25rem;
+}
+
+.wrap {
+  position: relative;
+  display: flex;
+  align-items: center;
+  border: none;
+}
+
+.wrap svg {
+  position: absolute;
+  left: 10px;
+  width: 24px;
+  height: 24px;
+}
+
+.input {
+  background-color: transparent;
+  font-weight: bold;
+  padding-left: 48px;
+  border: none;
+}
+
+.searchButton {
+  border: 0px;
+  width: 80px;
+  padding: 0.5rem 1rem;
+  border-radius: 0.5rem;
+  color: rgb(255, 255, 255);
+  background-color: rgb(24, 119, 242);
+}
